src/main/java/switchfive/project/dataModel/dataJPA/ResourceJPA.java
package switchfive.project.dataModel.dataJPA;

import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;

@Entity
@Table(name = "resources")
public class ResourceJPA implements JPA {

    @Id
    private String resourceID;

    private String userID;
    private String projectCode;
    private String startDate;
    private String endDate;
    private Double costPerHour;
    private Double percentageOfAllocation;
    private String role;
    private String email;

    public ResourceJPA() {
    }

    public String getResourceID() {
        return resourceID;
    }

    public void setResourceID(String resourceID) {
        this.resourceID = resourceID;
    }

    public String getUserID() {
        return userID;
    }

    public void setUserID(String userID) {
        this.userID = userID;
    }

    public String getProjectCode() {
        return projectCode;
    }

    public void setProjectCode(String projectCode) {
        this.projectCode = projectCode;
    }

    public String getStartDate() {
        return startDate;
    }

    public void setStartDate(String startDate) {
        this.startDate = startDate;
    }

    public String getEndDate() {
        return endDate;
    }

    public void setEndDate(String endDate) {
        this.endDate = endDate;
    }

    public Double getCostPerHour() {
        return costPerHour;
    }

    public void setCostPerHour(Double costPerHour) {
        this.costPerHour = costPerHour;
    }

    public Double getPercentageOfAllocation() {
        return percentageOfAllocation;
    }

    public void setPercentageOfAllocation(Double percentageOfAllocation) {
        this.percentageOfAllocation = percentageOfAllocation;
    }

    public String getRole() {
        return role;
    }

    public void setRole(String role) {
        this.role = role;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }
}
